basic controllers done for cities functionality

+++ src/controllers/cities.ts
@@ -9,6 +9,7 @@ import { ICityDetails } from '../models/interfaces/ICityDetails';
 import { MappedErrors } from '../utils/mappedErrors';
 import { TravelBrainError } from '../utils/travelBrainError';
 import { cityVisitsController } from './cityvisits';
+import { CityVisit } from '../models/cityVisit';
 
 /**
  * @class CitiesController
@@ -36,7 +37,7 @@ class CitiesController {
 			if (findErr) {
 				loogger.error(findErr);
 				const findError = new TravelBrainError(MappedErrors.MONGO.FIND_ERROR, {
-					mess: 'Unable to find city record'
+					mess: `Unable to find city record with id ${cityId}.`
 				});
 				cb(findError, { foundCity: 'nope' });
 			} else {
@@ -52,11 +53,11 @@ class CitiesController {
 			name: cityDetails.name,
 			country: cityDetails.country,
 			state: cityDetails.state || null,
-			mostRecentVisit: cityDetails.endDate,
+			mostRecentVisit: null,
 			numRestaurantsEaten: 0,
 			numSightsSeen: 0
 		};
-		City.create(cityRecord, (cityErr, insertedCityRecord): void => {
+		City.create(cityRecord, (cityErr, record): void => {
 			if (cityErr) {
 				loogger.error(cityErr);
 				const insertionError = new TravelBrainError(MappedErrors.MONGO.INSERTION_ERROR, {
@@ -64,38 +65,72 @@ class CitiesController {
 				});
 				cb(insertionError, { insertedCity: 'nope' });
 			} else {
-				cityVisitsController.newVisit(insertedCityRecord._id, cityDetails, cb);
-				// do some city visit code here
+				cityVisitsController.newVisit(record._id, cityRecord.cityId, cityDetails, true, cb);
+				this.latestCityId += 1;
 			}
 		});
 	}
 
 	public changeCityDetails = (cityDetails: Object, cityId: number, cb: Function): void => {
-		console.log(cityDetails);
-		console.log(cityId);
-		cb(null, { updated: 'yeet' });
-	}
-
-	public wipeCityDetails = (cityId: number, cb: Function): void => {
-		City.deleteOne({ cityId }, (deleteErr): void => {
-			if (deleteErr) {
-				loogger.error('Unable to delete city record');
-				const deletionError = new TravelBrainError(MappedErrors.MONGO.DELETION_ERROR, {
-					mess: `Error wiping the city with id ${cityId} off the map`
+		City.replaceOne({ cityId }, cityDetails, (replaceErr, res): void => {
+			if (replaceErr) {
+				const updateError = new TravelBrainError(MappedErrors.MONGO.UPDATE_ERROR, {
+					mess: `Unable to update record for city with id ${cityId}.`
 				});
-				cb(deletionError, { deleted: 'Unable to delete city record' });
+				cb(updateError, { updatedCity: 'nope' });
 			} else {
-				// delete city visits here
+				cb(null, { updatedCity: 'success' });
 			}
 		});
 	}
 
+	public wipeCityDetails = (cityId: number, cb: Function): void => {
+		let deletedVisits = true;
+		CityVisit.deleteMany({ cityId }, (delErr): void => {
+			if (delErr) {
+				deletedVisits = false;
+				loogger.error(delErr);
+				const deletionError = new TravelBrainError(MappedErrors.MONGO.DELETION_ERROR, {
+					mess: `Unable to delete visits for city with id ${cityId}`
+				});
+				cb(deletionError, { deletedCity: 'nope' });
+			}
+		});
+		if (deletedVisits) {
+			City.deleteOne({ cityId }, (deleteErr): void => {
+				if (deleteErr) {
+					loogger.error('Unable to delete city record');
+					const deletionError = new TravelBrainError(MappedErrors.MONGO.DELETION_ERROR, {
+						mess: `Error wiping the city with id ${cityId} off the map`
+					});
+					cb(deletionError, { deletedCity: 'nope' });
+				} else {
+					cb(null, { deletedCity: 'success' });
+				}
+			});
+		}
+	}
+
 	public getAllCities = (cb: Function): void => {
-		cb(null, {});
+		City.find({}, (findErr, records): void => {
+			if (findErr) {
+				const findAllError = new TravelBrainError(MappedErrors.MONGO.FIND_ERROR, {
+					mess: 'Unable to get all city records'
+				});
+				cb(findAllError, { foundAllCities: 'nope' });
+			} else {
+				const cityRecords: ICity[] = [];
+				for (let i = 0; i < records.length; i += 1) {
+					const c = _.pick(records[i], this.cityFields);
+					cityRecords.push(c);
+				}
+				cb(null, cityRecords);
+			}
+		});
 	}
 
 	public getRecentCities = (cb: Function): void => {
-		City.find({ sort: { mostRecentVisit: -1 }, limit: 10 }, (findErr, records): void => {
+		City.find({}, null, { sort: { mostRecentVisit: -1 }, limit: 10 }, (findErr, records): void => {
 			if (findErr) {
 				loogger.error('Error trying to find 10 most recent cities');
 				const findError = new TravelBrainError(MappedErrors.MONGO.FIND_ERROR, {
@@ -113,9 +148,22 @@ class CitiesController {
 	}
 
 	public searchCities = (param: string, value: string, cb: Function): void => {
-		console.log(param);
-		console.log(value);
-		cb(null, {});
+		City.where(param).equals(value).exec((findErr, records): void => {
+			if (findErr) {
+				loogger.error('Error trying to search cities');
+				const findError = new TravelBrainError(MappedErrors.MONGO.FIND_ERROR, {
+					mess: 'Unable to search cities'
+				});
+				cb(findError, { found: 'nope' });
+			} else {
+				const cities: ICity[] = [];
+				for (let i = 0; i < records.length; i += 1) {
+					const c = _.pick(records[i], this.cityFields);
+					cities.push(c);
+				}
+				cb(null, { cities });
+			}
+		});
 	}
 }
 /**
